// src/app/core/models/formalize-processes/formalize-processes.model.ts
export interface IFormalizeProcesses {
  goodNumber?: number;
  eventId?: number;
  stage?: number;
  notaryIdterc?: number;
  lotId?: number;
  notaryCli?: string;
  numNotaryCli?: number;
  cityNotary?: string;
  date?: string;
  assignmentnotDate?: string;
  jobNumber?: string;
  writingDate?: string;
  writingNumber?: number;
  writingAntNumber?: number;
  writingAntDate?: string;
  origin?: string;
}